Complete rework of #Criteria
Since my idea for the criteria class was fundamentally flawed and could
not be made to work properly, I borrowed the idea from solution-2 to use
single "criterions" and pass their conditions to them as block
arguments. This also allows me to easlily combine them with #and, #or
and #not.